+++ pkg/lib/cockpit-core.js
