+++ test/static-code
@@ -69,10 +69,6 @@ if [ "${WITH_PARTIAL_TREE:-0}" = 0 ]; then
 
     test_typescript() {
         test -x node_modules/.bin/tsc -a -x /usr/bin/node || skip 'no tsc'
-	# https://github.com/microsoft/TypeScript/issues/30511
-	# We can't tell tsc to ignore the .d.ts in node_modules/ and check our
-	# own cockpit.d.ts, so we do two separate invocations as a workaround:
-	node_modules/.bin/tsc --typeRoots /dev/null --strict pkg/lib/cockpit.d.ts
 	node_modules/.bin/tsc --checkJs false --skipLibCheck
     }
 fi
